backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.3.0
yfinance==0.2.40
beautifulsoup4>=4.12.0
lxml>=5.0.0
pandas>=2.2.0
PyJWT>=2.8.0
python-multipart>=0.0.9
asyncpg>=0.29.0
passlib[bcrypt]>=1.7.4
bcrypt<5.0                 # passlib 1.7 is incompatible with bcrypt 5.x (72-byte limit regression)
apscheduler>=3.10.4
alembic>=1.13.0
anthropic>=0.30.0
mcp>=1.0.0

# Security + ops (added in M1/M4/M6/M7 pushes)
cryptography>=42.0.0       # Fernet PII column encryption
slowapi>=0.1.9             # Rate limiting on /auth/token + /signup
aiosmtplib>=3.0.0          # Async SMTP for verification emails
reportlab>=4.0.0           # Form 8949 / Schedule D tax PDFs
sentry-sdk[fastapi,sqlalchemy]>=2.0.0  # Error tracking with PII scrub
boto3>=1.34.0              # S3 backups + AWS Secrets Manager
pyotp>=2.9.0               # TOTP admin MFA
stripe>=10.0.0             # M8 billing (Checkout + Portal + webhooks)

# Test deps
pytest>=8.0.0
pytest-asyncio>=0.23.0
anyio[trio]>=4.0.0
